🥔✨ `Marketplace`: `DeliveryAreas` may be `Archived` (#2024)
* 🥔✨ `Marketplace`: `DeliveryAreas` may be `Archived`

- https://github.com/zinc-collective/convene/issues/2014

This adds the `discard` gem so that we can archive objects. It also
updates the workflow for removing DeliveryArea to have an `Archive` step
first.

Once a `DeliveryArea` is archived, it can not yet be restored; but it
can be removed if there are no `Orders` associated with the
`DeliveryArea`

* 🧹 `Marketplace`: `DeliveryArea` Strip TurboStreams (#2026)

Now that we have a better idea how to use TurboFrames, the Streams seem
pretty unnecessary.

This also makes it so following links within the
`Marketplace::ManagementComponent` update the browser URL.

* ✨ `Marketplace`: Put Archived `DeliveryAreas` behind the `Archive` link (#2027)

* 🧹 `Marketplace`: `DeliveryArea` Strip TurboStreams

Now that we have a better idea how to use TurboFrames, the Streams seem
pretty unnecessary.

This also makes it so following links within the
`Marketplace::ManagementComponent` update the browser URL.

* ✨ `Marketplace`: Put Archived DeliveryAreas behind the `Archive` link

* One-line disable


* whitespace is hard


* don't not


* Better words

* Pull discardability/destroyability up to the Model

* Fix system test

---------

+++ app/furniture/marketplace/delivery_area_component.rb
@@ -28,20 +28,29 @@ class Marketplace
       delivery_area.persisted? && policy(delivery_area).edit?
     end
 
+    def discard_button
+      return unless discard_button?
+
+      ButtonComponent.new(label: "#{t("icons.discard")} #{t("discard.link_to")}",
+        title: t("marketplace.delivery_areas.discard.link_to", name: delivery_area.label),
+        href: delivery_area.location, method: :delete, scheme: :secondary)
+    end
+
+    def discard_button?
+      delivery_area.discardable? && policy(delivery_area).destroy?
+    end
+
     def destroy_button
       return unless destroy_button?
 
       ButtonComponent.new(label: "#{t("icons.destroy")} #{t("destroy.link_to")}",
         title: t("marketplace.delivery_areas.destroy.link_to", name: delivery_area.label),
-        href: delivery_area.location, turbo_stream: true,
-        method: :delete,
-        confirm: t("destroy.confirm"),
+        href: delivery_area.location, method: :delete, confirm: t("destroy.confirm"),
         scheme: :secondary)
     end
 
     def destroy_button?
-      delivery_area.persisted? && policy(delivery_area).destroy? &&
-        delivery_area.orders.empty?
+      delivery_area.destroyable? && policy(delivery_area).destroy?
     end
   end
 end
